moved movies into the components folder
